requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.4
python-dotenv